+++ NineChronicles.Headless/GraphQLService.cs
@@ -168,9 +168,10 @@ namespace NineChronicles.Headless
 
                 // Capture requests
                 Dictionary<string, HashSet<Address>> ipSignerList = new();
-                app.UseMiddleware<HttpCaptureMiddleware>(
+                app.UseMiddleware<HttpMultiAccountManagementMiddleware>(
                     StandaloneContext,
                     ipSignerList);
+                app.UseMiddleware<HttpCaptureMiddleware>();
 
                 app.UseMiddleware<LocalAuthenticationMiddleware>();
                 if (Configuration[NoCorsKey] is null)
